Bug 1917056 Part 2 - Move the special logic for table frame out of IntrinsicForAxis(). r=layout-reviewers,emilio
In the old logic, the table's content inline size is stored in the `min`
variable, and passed into `AddIntrinsicSizeOffset()` to account for margin,
border, and padding.

We don't seem to have test coverage for this behavior, so I added some WPTs in
the next part.

This patch shouldn't change behavior.

Differential Revision: https://phabricator.services.mozilla.com/D221236

UltraBlame original commit: ec7ee6605016173aa975b5dfcac864027e4cdb17