Bug 1917132: When clearing a RemoteAccessible's parent, set its cached IndexInParent to -1. r=nlapre
We set the cached IndexInParent to -1 in RemoveChild, among other places.
However, when a RemoteAccessible is moved, we clear its children array without calling RemoveChild (in DocAccessibleParent::ShutdownOrPrepareForMove).
We already call SetParent(nullptr) on the children to clear the parent, so set the cached IndexInParent to -1 in SetParent when the supplied new parent is null.

Differential Revision: https://phabricator.services.mozilla.com/D221288

UltraBlame original commit: 027ddef0767b43fa1f76ea7e56c1e04c64c27a88