Bug 1908639 - part 3 - Set no clipping by default when applying EngineViewClippingBehavior r=android-reviewers,Roger
This is useful in the scenario of coming back to browsing a 100vh page after finishing
a find in page session.
When using a navbar this is happening by default since in the same scenario the navbar
is reset thus triggering an initial clipping evaluation in the appropriate behavior.

Differential Revision: https://phabricator.services.mozilla.com/D220282

UltraBlame original commit: 446f6d4fcb59cc0fb99c4e639fd668020c8ef1ca