Bug 1916771 - [devtools] Set ViewportSize highlighter text color to avoid contrast issue in dark mode. r=devtools-reviewers,ochameau.
Since we weren't setting a color, the default one was used,
which means a light one when the page is in dark mode.
But since we draw the ViewportSize on a light rectangle already,
this was making giving the text a poor contrast.

Differential Revision: https://phabricator.services.mozilla.com/D221108

UltraBlame original commit: f14a4fb44c4fa5f0b0eb80f7af5e4ffd7a58fb5b