Bug 1914570 - Use the SubFormat field to determine sample format when the format code is WAVE_FORMAT_EXTENSIBLE. r=media-playback-reviewers,karlt
This refactors the method that fetches the layout from the same struct, and uses
it to fetch the SubFormat (that comes after the layout).

Spec: https://www.mmsp.ece.mcgill.ca/Documents/AudioFormats/WAVE/WAVE.html

but it's clearer here:
https://learn.microsoft.com/fr-fr/previous-versions/dd757713(v=vs.85)
https://learn.microsoft.com/fr-fr/windows/win32/api/mmreg/ns-mmreg-waveformatextensible?redirectedfrom=MSDN

The first two bytes of the SubFormat (that is a GUID) is the format we're
looking for.

Differential Revision: https://phabricator.services.mozilla.com/D220433

UltraBlame original commit: 86f6571b38723492f5fa2d7cce33208d6b8240b5