Bug 1166408 - Part 1: Rename BindName to BindUnqualifiedName and add new BindName. r=spidermonkey-reviewers,jandem
Rename the existing `BindName` to `BindUnqualifiedName` and add a new `BindName`
operation which doesn't stop at the first variables object.

It's a bit unfortunate having to rename the existing `BindName`, but it seems
like the cleanest solution:
- `BindUnqualifiedName` communicates more clearly that the lookup stops at a
  variables object, cf. `LookupNameUnqualified`.
- And plain `BindName` without any extra qualifiers follows the complete
  environment chain until the global object has been reached.

The implementation of the new `BindName` is quite similar to `BindUnqualifiedName`.
Notable differences:
1. It calls `LookupNameWithGlobalDefault` instead of `LookupNameUnqualified`.
2. `LookupNameWithGlobalDefault` doesn't need to check for `const` bindings
   and we don't have to create `RuntimeLexicalErrorObject`.

`BindNameIRGenerator` needs to be updated to match to the above two differences.

NOTE:
`LookupNameWithGlobalDefault` is currently also called for the `ImplicitThis`
byte code, but this will change in part 2.

Differential Revision: https://phabricator.services.mozilla.com/D221665

UltraBlame original commit: 2e114a523a56ca4c35e8b1ccfa6346aecde2fd14