Bug 1914191 - Part 3: Switch WPT canvas tests to using the unusedGraceMs pref, r=smaug
This test suite was previously using the keepProcessesAlive pref to reduce
process churn, and is the original motivation behind the new change. As there
are a number of test failures if we turn on the pref for all test suties, only
turn it on for this specific test suite to allow the change to land sooner.

Differential Revision: https://phabricator.services.mozilla.com/D221383

UltraBlame original commit: 088ced7f3f05068e89fedc1018c34a57f977f868