Bug 1903134 - Make parent PortMessage and PortDisconnect messages not be queries. r=extension-reviewers,robwu
This adds a new flag, `query`, to the _cast() argument `arg`. If this is true,
you get the current behavior where the messages sent to the child are queries,
and we return a promise that waits on one or all replies. Otherwise the
messages are all normal async messages and we return undefined.

This new flag is set to true for RuntimeMessage and PortConnect, and to false
for PortMessage and PortDisconnect.

Differential Revision: https://phabricator.services.mozilla.com/D222384

UltraBlame original commit: 123394f54e89591e011cb8a4b751e4a2df1ce8ff