Bug 1920008 - Fix swipeToSwitchTabTest UI test on API 34 r=ohorvath
The UI test was failing on API 34 when performing the swipe action on the toolbar because this AVD uses "gestures" for it's navigation.

To overcome this problem, I've created enableOrDisableBackGestureNavigationOnDevice that allows us to enable or disable the back gesture from the edge of the screen on the device.

The UI test:

Successfully passed 25x on API 30 ✅
Successfully passed 50x on API 34 ✅

Differential Revision: https://phabricator.services.mozilla.com/D222914

UltraBlame original commit: fb30f015c9fa10ad7408c0e89123fbf69516db2a