Bug 1913312 - Enable Persisted Search for non-default app provided search engines - r=adw,urlbar-reviewers
There's two main changes:

- Change `getSearchTermIfDefaultSerpUri` to enable it work for any app-provided engine.
- Show the correct Search Chicklet / Unified Search Mode Chicklet when the user loads a non-default SERP. Recall that Persisted Search is enabled in Nightly independent of Scotch Bonnet.

**Issue 1:**

 `this.searchMode` is not always known or consistent with the non-default search page when `setURI` is called.

Some scenarios:

- Clear search mode, switch tab, return to the original. The absence of search mode seems acceptable because the user was potentially in the middle of modifying their search or navigating to a URL.
- Clear search mode and reload a non-default SERP.
- Switch search mode, but reload non-default SERP.
- Go back/forward in tabhistory to return to a non-default SERP.

In the latter three cases, if the search terms were to persist again, I'd expect the most accurate search mode to be present.

**Issue 2:**

`userTypedValue` has been used as an indicator for whether or not search terms should persist. However, entering search mode causes userTypedValue to have a value, so that it can be restored when switching back to the current tab and across sessions. Thus, on blur, I have a slightly more complicated check to ensure the revert button is consistent regardless of whether the engine is default or non-default.

**Issue 3:**

Lastly, it's possible for the revert button to show up when switching from tab to tab while the address bar is focused. I added an additional check in this patch since it's a small addition, but I can move it out if it overcomplicates things.

I'm not 100% confident my solution is ideal since I ended up having to add hacky awaits in `browser_UrlbarInput_searchTerms_searchMode.js` for the search mode to appear...

Differential Revision: https://phabricator.services.mozilla.com/D222846

UltraBlame original commit: 013944963631e00ed789d7c1d89a01fc8e22ad0c