Bug 1919927 - Implement `contenteditable="plaintext-only"` in the DOM module r=smaug
It requires a lot of changes in the editor module, and the layout module may
need `read-write-plaintext-only` and `user-modify: read-write-plaintext-only`.
When we work on the modules, we need the basic support under `dom/`.  This
patch adds `plaintext-only` support only when the new pref is enabled.

Once the pref is enabled, the following failures will be fixed:
* https://searchfox.org/mozilla-central/rev/f2c181a7ab3bfea4d2266521e6eac713630479b3/testing/web-platform/meta/html/editing/editing-0/contenteditable/contenteditable-enumerated-ascii-case-insensitive.html.ini#4-5,7-8,10-11,13-14
* https://searchfox.org/mozilla-central/rev/f2c181a7ab3bfea4d2266521e6eac713630479b3/testing/web-platform/meta/html/editing/editing-0/contenteditable/user-interaction-editing-contenteditable.html.ini#4-5,7-8

FYI: `Element::IsContentEditablePlainTextOnly()` is useful for `HTMLEditor`
to check whether active editing host is so or not.  Therefore, I'd like to
add it even though there is no user of it.

Differential Revision: https://phabricator.services.mozilla.com/D222900

UltraBlame original commit: 0f470e1a67afe6044b03ba5c8dd92daa3037de8a