Bug 1916363 - [devtools] Adapt request list icons in High Contrast Mode. r=devtools-reviewers,ochameau.
This sets HCM variable for icon colors that weren't already overridden
in HCM (error, warning). Since those colors are used for icons that
are already distinct from each other, this doesn't cause any issues.

The slow icon needed a bit more work as it's an absolutely positioned element that
can overlap the text. In regular mode, a gradient is added on its left side so
the icon has a solid background and the text slowly fades before the icon.
In High Contrast Mode, the background image is removed, so the icon was fully
overlapping the text, making it unreadable.
To fix this, we set a plain solid color on the icon in HCM, which also needed
some dimensions tweaks to make it look nice.

Differential Revision: https://phabricator.services.mozilla.com/D222247

UltraBlame original commit: 2a0022d78bca8c5f0d3ac4f01f1c35dd7b22d8c7