Bug 1921001 - [devtools] Make markup view filter input clear button focusable with the keyboard. r=devtools-reviewers,ochameau.
This simply removed the negative tabindex attribute that was set on it.
We take this as an opportunity to automatically focus the input when
the clear button is used.
A test case is added to cover those changes.

Differential Revision: https://phabricator.services.mozilla.com/D223589

UltraBlame original commit: 3affd8c29527aff5ca2a773aff6747f36fb732fc