Bug 1919683 correct documentation of media.ruin-av-sync.enabled r=media-playback-reviewers,padenot
The documentation implied that a minimum queue size of 0 would remedy A/V sync
but this actually just changes the effect to a playback stall, as reported in
bug 1316571 and others.

A subsequent patch will restore the minimum queue size of 1, but still
drop frames while the video is playing.

Differential Revision: https://phabricator.services.mozilla.com/D223693

UltraBlame original commit: 93492262ee8546303389e03cd177b50a95757b9a