Bug 1921026 - Update engine view's bottom margin whenever the fixed navbar is shown or hidden r=android-reviewers,skhan
Whenever the tab is configured with a fixed toolbar the engine view has it's bottom margin
set to the height of the bottom toolbar to ensure it it shown on top of it and no web content
is clipped.
When interacting with an `input` element in a webpage the addressbar remains shown so there
is no issue but the navbar gets hidden so the engine view's bottom margin has to be updated
to account for the new layout.

Differential Revision: https://phabricator.services.mozilla.com/D224129

UltraBlame original commit: 6dafb6e4cb913ad9fc10924057df2955b18a40cb