Bug 1921404 - Fix download tests affected by notification stack difference on API 34 r=ohorvath
The system notification is different on API 34 and because of this there were problems when trying to expand it.
To be able to expand the notification we need to perform a swipe down action on the notification.
Only after expanding it we'll be able to see the app's name.
Added some conditions to make them work in both API 30 and 34 and I had to link systemNotificationCantBeDismissedWhileInProgressTest to another test (specific behaviour for Android 14) refactor and make it run only on API 34.



  - All 3 UI tests successfully passed on API 30 25x (before Oana's [[ https://phabricator.services.mozilla.com/D223705 | change ]] that landed yesterday ) ✅
  - All 3 UI tests successfully passed on API 34✅

Differential Revision: https://phabricator.services.mozilla.com/D223916

UltraBlame original commit: 3883c657f3aadbf817b74871a0bd47f2879c9bc4