Bug 1921845 [wpt PR 48387] - Revise the behavior of wpt-tests when the embed element is not displayed, a=testonly
Automatic update from web-platform-tests
Revise the behavior of wpt-tests when the embed element is not displayed

According to the spec[1], the `<embed>` element is said to be
potentially active when is being rendered.
> The element is being rendered, or was being rendered the last time
> the event loop reached step 1.
To ensure that the `<embed>` element does not fetch request when it is
not displayed, include the embed-represent-nothing-04.html test case.

According to the spec[2], the `<object>` element that is not rendered
will not fetch request.
> if the element is not being rendered, then jump to the step below
> labeled fallback.
> Fallback: The object element represents the element's children.
In the `embed-in-object-fallback-2.html` test case, the `<object>`
and `<embed>` element are not displayed, so no elements
should be loaded. Remove `display: none` in this test case to
test the `<embed>` element nesting inside the `<object>` element.

`window.onload` is triggered  until the embed object is finished
loading. `embed-document-under-content-visibility-*` not determine
whether the `<embed>` element is loaded.

[1]: https://html.spec.whatwg.org/multipage/#the-embed-element
[2]: https://html.spec.whatwg.org/multipage/#the-object-element

Bug: 5877938
Change-Id: I722f86098aef108e39637da4ba79d7588c56b7cd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5898095
Reviewed-by: Mason Freed <masonfchromium.org>
Reviewed-by: Traian Captan <tcaptanchromium.org>
Commit-Queue: Peng Zhou <zhoupeng.1996bytedance.com>
Cr-Commit-Position: refs/heads/main{#1361889}

--

wpt-commits: 03e9faee0674e595d84577a1cf624a8da0cc6a0e
wpt-pr: 48387

UltraBlame original commit: 5a0edc423f56dc675d446fa4700ce8144c7898c7