Bug 1916242 - Add an option in the secret settings to keep the debug menu enabled after restart. r=007,android-reviewers
This patch adds a new pref controlled by a secret setting. When not set
or set to false, nothing changes, i.e. the debug menu is enabled for the
duration of the session via the usual trigger (multiple clicks on the
Firefox logo).

When the pref is set to true, the debug menu is enabled by default.

Differential Revision: https://phabricator.services.mozilla.com/D223944

UltraBlame original commit: f1acac810cc08742c16884a9a6e070a844de1eef