Bug 1915863 part 4 - Simplify control flow in processBundle a bit. r=jseward
The `canAllocate` variable is set before the loop and is never mutated after that
point. This means it's simpler to check it outside the loop instead of on each
iteration.

This also lets us move the `chooseBundleSplit` call to the end of the function.

Differential Revision: https://phabricator.services.mozilla.com/D224910

UltraBlame original commit: 57689ae8d79db1a1c2a4a7706b9c0e164d07afd2