Bug 1922126 [wpt PR 48416] - [flex] Avoid layout when determining the intrinsic block-size., a=testonly
Automatic update from web-platform-tests
[flex] Avoid layout when determining the intrinsic block-size.

https://chromium-review.googlesource.com/c/chromium/src/+/5749947
Did some changes to correctly call kContent & kIntrinsic within flex
for the purposes of determining flex-basis.

This uncovered a bug where we'd layout a flex-item (with an
aspect-ratio, and a container-query container), and not correctly
relayout when the size changed.

This is the first part of the mitigation. E.g. when we have size
containment we don't need to run layout to determine the intrinsic
block-size.

There is a set of other issues in this area. But this mitigates the
regression for the moment.

Bug: 370480725
Change-Id: I8301a424fca86392c2ec1a7cfcf867142e413b31
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5901118
Commit-Queue: Ian Kilpatrick <ikilpatrickchromium.org>
Reviewed-by: David Grogan <dgroganchromium.org>
Reviewed-by: David Baron <dbaronchromium.org>
Cr-Commit-Position: refs/heads/main{#1363322}

--

wpt-commits: 5e053f07ddac19076db934f89d9d355b49a75002
wpt-pr: 48416

UltraBlame original commit: 13737f1a32ea9c0d4cede7aebb2bfa77f0578b93