Bug 1923199: Correctly report OOMs when allocating structs. r=rhunt
We were simply not performing the relevant check after performing an
instance call to allocate a struct. This patch implements this check but
also refactors a bit to centralize our handling of these FailureMode
checks, and to thread a wasm::BytecodeOffset through to the point of
failure.

Differential Revision: https://phabricator.services.mozilla.com/D225296

UltraBlame original commit: bbd21193acc27124d8af5ba9951dda36a9e80265