Bug 1924504 - Make GCRuntime::getOrAllocChunk leave chunk on the empty chunks list r=sfink
The problem is that we can fail to allocate an arena in a new chunk for various
reason, but currently we eagerly move the chunk from the empty to the available
chunks list when getting an empty chunk and this causes an assertion failure
later when the chunk is found to stil be empty.

The patch makes it so that we leave the chunk on the empty chunk list until we
successfully allocate an arena in it. It also adds more assertions during the
decommit process.

I don't think anything here could have caused a crash.

Differential Revision: https://phabricator.services.mozilla.com/D225509

UltraBlame original commit: c63801497940b667d2c47373aee95312502a2acb