Bug 1923899 - [bidi] Wait for currentWindowGlobal when sending commands to window global message handlers r=webdriver-reviewers,Sasha
When attempting to send commands on a very early browsing context, the command might fail
due to a missing currentWindowGlobal.
This patch adds a helper to wait for the currentWindowGlobal property to be set and uses it from
RootTransport in order to avoid this issue.
A browser mochitest is added to cover this scenario.

Differential Revision: https://phabricator.services.mozilla.com/D225281

UltraBlame original commit: 879b49c5132b578dc0da57fc0aa6aa633c9b2f24