Bug 1911021 - Add an API tp CycleCollectedJSContext to enable/disable tracing r=smaug
Basically, in order to allow turning on JS execution tracing from a convenient place
within the profiler, we want a threadsafe endpoint hanging off of the
`CycleCollectedJSContext`. The profiler holds onto a pointer to the `CycleCollectedJSContext`
controlled via `PROFILER_SET|CLEAR_JS_CONTEXT` and calls this from a background thread
when profiling starts, or from the cx's own thread if the profiler is already running
when it is registered.

Differential Revision: https://phabricator.services.mozilla.com/D222755

UltraBlame original commit: 94505220c573791e05e2ce1b82adcb68d81567fc