Bug 1922926 part 1 - Change SubstringKernel to return the rope if the substring has the same length. r=sfink
This case is pretty common and we should optimize it to not flatten the rope
or allocate new strings.

This also matches the Ion inline code for `LSubstr` because it does the length
check before checking if the string is a rope.

A full Speedometer 3 run has only a few hundred calls to `SubstringKernel` with a
rope, but more than 95% of them are for this case.

Differential Revision: https://phabricator.services.mozilla.com/D225618

UltraBlame original commit: 84755be14a1b3c27884d769bbd3d7098b2bf7fe3