Bug 1805367 - Prevent text input from overlapping scrollbar in tree widget. r=dholbert,desktop-theme-reviewers,reusable-components-reviewers,dao,mstriemer
By setting `display:none` on the collapsed scrollbar, we can more
accurately calculate the `width` and `left` properties needed for the
text input in the XUL tree element.

By implementing the previously mentioned `display:none` change, we have
to update the calculations for the various properties on the text input.
Using the new `textPadding` variable, which is the amount of space from
the physical left wall of the cell to the start of the text in the cell,
we more calculate the `left` and `width` for the text input.

Since we are using `left` to absolutely position the text element, the
RTL calculation needs to be the cellRect.x plus text padding plus the
scrollbar width. However, since cellRect.x returns a negative value in
RTL (since getCoordsForCellItem does not give us the correct value
in this mode, see bug 1708159), we clamp this value to 0. This is
because the text input is absolutely positioned relative to the tree
stack element, having a negative x-coordinate would put the text input
outside the tree's boundaries. Additionally, if we're editing a cell
that is not in the first column, we need to add the scrollbar width
again to offset the text input. This is because the leftmost column
does not shrink when the vertical scrollbar is present.

In the LTR mode, we calculate `left` the same way without adding the
scrollbar width. Additionally, if the cell that is being edited is in
the last column, we remove the scrollbar width from the `width`
property. Otherwise, the text input would overflow past the tree's
boundaries. This is because consumers of getCoordsForCellItem have to
handle potential overflow and scrollbar overlapping issues.

These changes affects both LTR and RTL locales and prevents overflow in
both locales.

Differential Revision: https://phabricator.services.mozilla.com/D225558

UltraBlame original commit: 2e9eada4753d36c564167d7758db2ea2214cc2d1