Bug 1917652 Part 4: Update IPC timing expectations for gpu process platforms. r=mconley,taskgraph-reviewers,jmaher
The message timings that were modified in this test are sent by the
GPUProcessManager, and this test was using WIN and !WIN as a proxy for
whether or not a GPU process was in place. The modifications here make
that more explicit, although they remain approximate. Anything that is
conditioned on GPUPROCESS needs to "ignoreIfUnused" because those
markers will be sent or not sent depending on the presence of the GPU
process, which has complex controls not completely covered by the logic
in this test.

Differential Revision: https://phabricator.services.mozilla.com/D225565

UltraBlame original commit: 483c42666d63d12055535e2cb5e22ca3684a0019