Bug 1917201: Remove cancel button from modal while waiting for user to respond to Windows geolocation dialog r=Gijs
The cancel button is not really useful here and is probably confusing some
users, since they are pressing it anyway, and it is causing this crash. The
ideal fix would be for the Windows dialog to be modal instead of just immovable,
but it isn't.

Differential Revision: https://phabricator.services.mozilla.com/D225154

UltraBlame original commit: 0bbd77193d50a7838e233f0bdb80e3708e7c9c42