Bug 1925130 [wpt PR 48654] - Remove TextCluster caching in TextMetrics, a=testonly
Automatic update from web-platform-tests
Remove TextCluster caching in TextMetrics

In a previous CL, the getTextClusters() API was implemented with a cache
for the created TextCluster objects. This API may include as a parameter
a dictionary with text align and baseline options. Since the cache is
used if it was already calculated, subsequent calls with different
options return clusters in the wrong positions.

This CL removes this cache to guarantee that the results are always
calculated from scratch taking into account the current options passed

Tests were added using the CanvasTest font to check that the positions
are coherent with the options passed.

A future CL with reincorporate this cache in a way that allows
offsetting the stored clusters following the current align and
baseline options passed.

Bug: 341213359
Change-Id: I4cabdb62d05a9c8763300c0d268d7794ca9f160b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5933690
Reviewed-by: Jean-Philippe Gravel <jpgravelchromium.org>
Commit-Queue: Andres Ricardo Perez <andresrperezchromium.org>
Cr-Commit-Position: refs/heads/main{#1369650}

--

wpt-commits: ed58c696708173f4f7494a501eec29756b116036
wpt-pr: 48654

UltraBlame original commit: dbf91baafef2e67c1d4e935053cbe54008920e03