Bug 1920821 - Enable recording with mitmproxy 11 on Win/Mac. r=perftest-reviewers,afinder
This patch enables the ability to record pageload tests on platforms on
Win/Mac. Previously this was only available on Linux. Until the ubuntu
version on the linux platforms are updated, that is restricted to mitm8
recordings, meanwhile 11 can be used for win/mac.

Differential Revision: https://phabricator.services.mozilla.com/D225362

UltraBlame original commit: 0781de425698668c4d22cb127a7022122b26bb36