Bug 1911457 - Do a reflow on the root scroll container when the horizontal overflow area is changed to recompute minimum-scale size. r=layout-reviewers,emilio
On mobile environments, the `minimum-scale` size, which allows users to zoom
out to scales less than 1.0, should include the ICB's overflow, even when
`overflow-x:hidden` is specified on the root element.

This change ensures that we update the `minimum-scale` size when the overflow
of the ICB changes.

Differential Revision: https://phabricator.services.mozilla.com/D226416

UltraBlame original commit: 3033cffb3c50832accbfb5d8ffaaddc9247b7080