Bug 1880582: Make nsCocoaWindow hold a death grip on its native window until its destructor. r=mstange
This is a simplified version of a fix authored by Steven Michaud. This
creates a death grip in DestroyNativeWindow() and holds it until the
nsCocoaWindow is destroyed. This seems to satisfy the various run loops
in macOS which might invoke DestroyNativeWindow() without holding a
reference to the window itself.

Differential Revision: https://phabricator.services.mozilla.com/D226169

UltraBlame original commit: 3a818b2e1d6d65d590156d06a06906112a5283fe