Bug 1926668 - Support ExtendMode::REPEAT in DrawTargetWebgl. r=aosmond
If, by default, all DrawTargetWebgl textures are created with ExtendMode::REPEAT,
then we can support a hardware-accelerated repeat mode for free with no shader
alterations.

The shader already implements clamp to edge behavior manually to support texture
atlasing, so to allow textures to repeat, we just have to force them to be a
standalone texture and set the clamping to wider bounds to not interfere with
the hardware repeat.

Differential Revision: https://phabricator.services.mozilla.com/D226703

UltraBlame original commit: d999e2f0ea60e16b482bf43505bac17634d09a36