Bug 1926809 - Make inline-block-slice-7 reftest stable by removing usage of viewport-relative units. r=dholbert
Change all the viewport-relative units to inch so that it is more stable in
reftest-paged environment. In addition, add `margin-block-start` to `.after` in
the reference so that we can use `==` rather than `!=` to compare the test.

Differential Revision: https://phabricator.services.mozilla.com/D226959

UltraBlame original commit: e47c47c1231f63fa172f771bd2b7f19d105ee911