Bug 1927114 - Rename nsLineBox::RIndexOf to RLIndexOf and make it search from both ends of the line. r=dshin
This does not change the result, only the performance characteristics:
the search will be fast if the target frame is near either edge of the line,
not only if it's at the end.

Also includes a bit of trivial cleanup that I noticed in the surrounding
nsLineBox code. (No changes to behavior.)

Differential Revision: https://phabricator.services.mozilla.com/D226916

UltraBlame original commit: 4cb0ee10f36015521a86eb1ff21d536b4d9c72bd