Bug 1916391 - [devtools] Display Netmonitor Statistics charts legend in High Contrast Mode. r=devtools-reviewers,bomsy.
The colors that are defined for the charts are specified in High Contrast Mode,
so we can display them in the legend just fine.
This requires making the legend element their own element, as we don't want to
opt out of forced colors for the cell content.

Differential Revision: https://phabricator.services.mozilla.com/D226788

UltraBlame original commit: 6967e4d5d89de55c3fe5fee4904a8828f195b6d2