Bug 1926763 - Make the zombie process warning message debug-only. r=ipc-reviewers,mccr8
The situation that the zombie check from bug 1881386 was added for is
applicable to our CI but not to normal use.  We do sometimes detect
zombie processes in normal operation, but this isn't inherently a
problem.  Therefore, the log message for this is made debug-only to avoid
confusion for users of release builds.

Also, a comment added in bug 1881386 was missing the bug number; this has
been fixed.

Differential Revision: https://phabricator.services.mozilla.com/D226863

UltraBlame original commit: fe26eba04a72d454abb405a9c86a5f313c42bbc8