Bug 1926983 - [devtools] Don't scroll markup node into view on click. r=devtools-reviewers,jdescottes.
Setting the focus on the markup element was also triggering
the element to be scrolled into view, which causes a scroll
by default.
We don't need to scroll when the user clicked on the element,
as it's already visible, and worse, could cause horizontal scroll
to be reset, making it hard to edit an attribute.

Differential Revision: https://phabricator.services.mozilla.com/D226877

UltraBlame original commit: 2b57170f090c15ad94afcd928cadbd81c7b014e7