Bug 1927597 - [ci] Use "base" task to determine whether to run `osint` variants in cron, r=taskgraph-reviewers,jmaher
In bug 1924447 I attempted to fix the os-integration cron task so that
it wouldn't run any tasks that wouldn't normally be run on central by
applying the standard_filter to it. However, this backfired and instead
filtered out *all* of the osint tasks from the cron. There were actually
two problematic filters here:

1. filter_out_cron *facepalm*
2. filter_for_project, because *all* os-integration variants have
   `run_on_project: []` set on purpose as we only want to run them in
   this cron.

So really what we need to do, is check to see whether the "base" task
(aka, the task the osint task is derived from) should be filtered out of
not. We find the "base" task by stripping `-osint` off the label, then
looking either for a label without chunks, or for a chunk one.

Differential Revision: https://phabricator.services.mozilla.com/D227096

UltraBlame original commit: c189f2d9b112a1c6f5aea66c95bfd41ea75b767c