Bug 1925007 - [tryselect] Use 'filters' parameter rather than 'target_tasks_method' for tryselect, r=taskgraph-reviewers,jmaher
It turns out that overriding the `target_tasks_method` in ./mach try
comes with a major disadvantage. Namely that when using the
`-p/--parameters` flag, any non-standard `target_tasks_method` defined
in there will also be overwritten. For example, this makes it impossible
to make `./mach try` target the set of tasks that would be selected by a
cron task.

Luckily Taskgraph has a concept of "filters" which is pretty much
identical to how `target_tasks_method` works (I think the intent was for
filters to replace target_tasks, but that never ended up happening).

This means we can make the `./mach try` methods sit outside of
`target_tasks` and not override the original `target_tasks_method`
unless we mean to.

This change should be a simple refactor and not change any
functionality.

Differential Revision: https://phabricator.services.mozilla.com/D226554

UltraBlame original commit: ec0ffbb31137a55baccf5caf7c8f8f565c9eefe6