Bug 1927776 - don't analyze native HTML format in Content Analysis r=dlp-reviewers,handyman
There's a mismatch in the formats that various calls to Content Analysis
want - kNativeHTMLMime is not in DataTransfer's kNonPlainTextExternalFormats
but it is in HTMLEditor::HTMLTransferablePreparer::AddDataFlavorsInBestOrder().
This was causing multiple dialogs to be shown.

In this case, I don't think checking kNativeHTMLMime in addition to kHTMLMime
really adds anything, so I'm removing it. This fixes this problem, and also
removes a pretty useless Content Analysis call to boot.

Differential Revision: https://phabricator.services.mozilla.com/D227231

UltraBlame original commit: de2db2f7330f747a3d8bc052ee2eb3b8a9a7a410