Bug 1928003 - [devtools] Avoid triggering a11y failure in browser_markup_tag_edit_long-classname.js. r=devtools-reviewers,ochameau.
The test was failing accessibility check because we're clicking
on the tag line. Since we're handling element selection just fine
with the keyboard, we don't need to have the line being a proper
target, so we can disable the check.

Differential Revision: https://phabricator.services.mozilla.com/D227800

UltraBlame original commit: 039f3b62957434962e8f4f3721941414aa389eeb