Bug 1927677 - part 1 - Remove divider from on top of the OS navigation bar r=android-reviewers,twhite
`initializeMicrosurveyPrompt` is called in scenarios when we won't also show the navbar.
We wanted to show dividers on top of the bottom toolbar.
In this case the divider would be shown below it with no UX reason so it can be avoided.

Differential Revision: https://phabricator.services.mozilla.com/D227355

UltraBlame original commit: 5d64beaf4b6186bea93a1af51ae807a9f4537a5f