Bug 1928151 - Improve alpenglow dark frame background. r=desktop-theme-reviewers,sclements,dao
With the previous patch, this color is used more directly (it's already
used in light mode, but in dark mode toolbar_background is almost opaque
which papers over it).

Make it the bottom of the gradient background, to improve the behavior
with the previous patch.

Differential Revision: https://phabricator.services.mozilla.com/D227586

UltraBlame original commit: 98b2b793622a16cd4c194b64710d55cdc34cbcf3