Bug 1929182. Allow prerender for the GenerateGlyphMask case. r=jwatt
GenerateGlyphMask is used for background-clip: text, of which our implementation is pretty buggy. In fact, the testcase of this bug we ar buggy on even before the regressing bug, the regressing bug just made it worse. So we already didn't handle this situation properly and we happened to disturb it and make it worse. Let's just restore the old behaviour for the unusual background-clip: text case, but we still keep the win in display list building for all other (much more common) cases.

Differential Revision: https://phabricator.services.mozilla.com/D227950

UltraBlame original commit: a0919cc4b424f63adc4ec34d8282e4dfba10190a