Bug 1918139: Enable perf jitdump integration by default for nightly.  r=iain
This patch enables access to jitdump and perfspewer in nightly builds so that it is easier to profile these builds.  It is only active if the IONPERF environment variable is set and should not have any visible impact on performance by default.  It will also add a new `const char* name_` member to the VMFunctionData class for nightly builds.

Differential Revision: https://phabricator.services.mozilla.com/D226644

UltraBlame original commit: eb055c6db9547a4d1727a95c8037d1426eecb99e