Bug 1929511: Refactor TryRegExpFlagFromChar r=dminor
In a weird case involving stack overflow, it's actually possible for this to be called with an invalid flag. I don't remember why I originally decide to assert against that, but it doesn't seem worthwhile.

While we're here, V8 replaced all uses of v8::base::Optional with std::optional, and this was the final remaining instance, so I've updated it.

I'm not including the testcase because it depends on stack overflow being triggered with extreme precision in native code.

Differential Revision: https://phabricator.services.mozilla.com/D228211

UltraBlame original commit: 958d228b8b001f6cdb8fac4a97e8b1355d8a4064