Bug 1922506: Implement CanvasRenderingContext2D::GetDebugInfo. r=tjr,jgilbert,webidl,smaug
For our data collection efforts, we want to be able to tell whether a canvas we are reading is HW accelerated or not. For that reason, we want to implement CanvasRenderingContext2D::GetDebugInfo. We are only really interested in isAccelerated property, but added 3 other properties that might be useful in the future, or for other developers.

Differential Revision: https://phabricator.services.mozilla.com/D224495

UltraBlame original commit: 8fb993ba9d49d906a2063e62dc633818fae6c111