Bug 1927113 - Ensure persisted search doesn't show search mode from another tab - r=adw,urlbar-reviewers
this.searchMode can return null even if a search mode is shown
in the address bar, as the getter uses the current browser as the
context and attempts to use the state data. I try to solve this
by relying on the fact that a tab switch and invalid pageproxystate
results in restoring the search mode state.

Differential Revision: https://phabricator.services.mozilla.com/D228483

UltraBlame original commit: a9c8f6767ee21c1096932aa2bfcc5b3428c9f16b