Bug 1911521 - [devtools] Wait for no pseudo rule to be displayed in browser_rules_pseudo-element_01.js r=nchevobbe,devtools-reviewers
The initial check for nb == 0 should not be needed now that we test that the expected number of elements are displayed. The initial version of the patch was only using `some` and there it made sense to special case nb == 0, but we ended up improving this before landing.

Differential Revision: https://phabricator.services.mozilla.com/D228660

UltraBlame original commit: 4bf7c0efa609c107dc012835a3740ad629c36638