Bug 1923208 - Store animation index into AnimationEventDispatcher. r=layout-reviewers,birtles,jwatt,emilio
We need this information when comparing the cancelled transitions/animations
in the following patches.

Note that we have to reorder the calling of `Animation::Cancel()` and
the update of `mAnimationIndex` to make sure we enqueue the cancel event
with the correct `mAnimationIndex`.

Differential Revision: https://phabricator.services.mozilla.com/D227110

UltraBlame original commit: f94eadf5885014930f7269a83707705a53d824cf