Bug 1930633. Change ViewportUtils::GetVisualToLayoutTransform to take an nsIContent instead of a ViewID. r=hiro
In a lot of cases we go from frame/content, look up the id, then GetVisualToLayoutTransform looks up the content from the id. This is a waste. We should just pass in a content, and the one case that starts with an ViewID can look up the content to pass in since we were going to look up the content anyways.

The hashtable lookup of the content/id is showing up as biggest chunk the work in the event handling display list build that we do to handle synth mouse move events during sp3. And synth mouse moves have been identified as something we do that stands out as extra work.

Differential Revision: https://phabricator.services.mozilla.com/D228653

UltraBlame original commit: 0225286d7688b94be86e4eb5a967c18ca9a3c953