Bug 1923208 - Fallback to compare the order of event targets for CSSTransition if they have the same time. r=layout-reviewers,firefox-animation-reviewers,emilio,birtles
Since the cancelled transition has no owning element and its transition
generation is out-of-date, so we have to compare the event targets and
the transition generations in `AnimationEventInfo`. We still follow the
composite order defined in the spec but use the event target and transition
generation before cancelling this transition, to get the correct event order
if we have the same schedule time.

Differential Revision: https://phabricator.services.mozilla.com/D225117

UltraBlame original commit: f7ab9fb6f6a9a21023c500185ef1c4e4bc1108db