Bug 1926226 - Use only one reserved slot for reflected attributes returning a FrozenArray. r=edgar
HTML reflected IDL attributes that return a FrozenArray<T>? need to cache the
FrozenArray. This switches them from using a reserved slot on the JS object
per attribute to using an array of values for all of these types of attributes
for a particular interface, which is then stored in one reserved slot on the JS
object. This reduces the memory reserved for these attributes if they're not
actually used.

Differential Revision: https://phabricator.services.mozilla.com/D228797

UltraBlame original commit: 5560b57840750f30bb4cbce4661a4d37c76e647b