Bug 1931778 - Remove unnecessary condition in PreXULSkeletonUI. r=win-reviewers,rkraesig
We always have Dwm{Get,Set}WindowAttribute because we fail if we can't
load them.

This in practice means that we release the DC and uncloak the window a
bit later (due to the MakeScopeExits), but I think that's fine.

Differential Revision: https://phabricator.services.mozilla.com/D229267

UltraBlame original commit: e2831635ed6577be250649069e88d46fb6841a0e